requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
pdf2image==1.17.0
pillow==10.3.0
Werkzeug==3.0.3
pdfplumber==0.5.28
PyMuPDF==1.20.2
